organize styles files & add calculator basic styles
