--- src/components/Button.scss
@@ -1,4 +1,0 @@
-button {
-  width: 40px;
-  height: 40px;
-}
