--- src/components/Calculator.scss
@@ -1,19 +1,0 @@
-.App {
-  height: 100vh;
-  background-color: antiquewhite;
-  display: flex;
-  flex-flow: column;
-  align-items: center;
-  justify-content: center;
-}
-
-.App-header {
-  font-size: calc(30px  +  2vmin);
-  font-weight: bold;
-  color: brown;
-  cursor: pointer;
-}
-
-.App-header:hover {
-  text-decoration: underline;
-}
