+++ src/index.js
@@ -1,7 +1,7 @@
+import './styles/reset.scss';
 import React from 'react';
 import ReactDOM from 'react-dom';
 import Calculator from './components/Calculator';
-import './index.css';
 
 const attachable = document.getElementById('root');
 ReactDOM.render(<Calculator />, attachable);
