+++ src/styles/variables.scss
@@ -1,0 +1,10 @@
+@import url('https://fonts.googleapis.com/css?family=Open+Sans:300,600&display=swap');
+
+
+// fonts
+$font-stack: 'Open Sans', sans-serif;
+
+// colors
+$primaryColor: #141414;
+$secondaryColor: #666666;
+$offwhite: #f5f5f5;
